Extracting Positions Tabs out from main Component
