+++ alca-app/src/components/index.js
@@ -6,6 +6,7 @@ export * from "./Debug";
 export * from "./HelpDropdown";
 export * from "./NavigationBar";
 export * from "./Page";
+export * from "./PositionsTabs";
 export * from "./SnackbarMessage";
 export * from "./SubNavigation";
 export * from "./Web3NetworkMenu";
